resolve build warning in alpine_packets.cpp

+++ game_patch/multi/alpine_packets.cpp
@@ -191,7 +191,7 @@ void af_send_damage_notify_packet(uint8_t player_id, float damage, bool died, rf
     damage_notify_packet.header.type = static_cast<uint8_t>(af_packet_type::af_damage_notify);
     damage_notify_packet.header.size = sizeof(damage_notify_packet) - sizeof(damage_notify_packet.header);
     damage_notify_packet.player_id = player_id;
-    int rounded_damage = std::round(damage);
+    int rounded_damage = static_cast<int>(std::round(damage));
     damage_notify_packet.damage = static_cast<uint16_t>(std::max(1, rounded_damage)); // round damage with minimum 1
 
     damage_notify_packet.flags = 0; // init flags
